+++ examples/resources/english.json
