+++ examples/resources/portuguese.json
