+++ examples/resources/spanish.json
